test: Make as many image conversion tests work as possible.
