+++ test/_test_data/images/.gitignore
@@ -1,0 +1,2 @@
+thumbs/
+tmp/
